// utils/types.ts
export enum StorageProtocolID {
  IPFS = 1,
}

export interface MetadataPointer {
  /**
   * The decentralized storage protocol
   * Read more here: https://github.com/gitcoinco/grants-round/blob/main/packages/contracts/docs/MetaPtrProtocol.md
   */
  protocol: StorageProtocolID
  /**
   * The identifier which represents the program metadata on a decentralized storage
   */
  pointer: string
}
// ...
export type ProjectRequirements = {
  twitter: {
    required: boolean
    verification: boolean
  }
  github: {
    required: boolean
    verification: boolean
  }
}

export interface ApplicationMetadata {
  questions?: SchemaQuestion[]
  requirements: ProjectRequirements
}

export interface Round {
  /**
   * The on-chain unique round ID
   */
  id?: string

  chainId?: number

  /**
   * Metadata of the Round to be stored off-chain
   */
  roundMetadata: {
    name: string
    programContractAddress: string
    roundType: string
    eligibility?: {
      description: string
      requirements: { requirement: string }[]
    }
    quadraticFundingConfig: {
      matchingFundsAvailable: number
      matchingCap: boolean
      matchingCapAmount?: number
      minDonationThreshold?: boolean
      minDonationThresholdAmount?: number
      sybilDefense?: boolean
    }
    support?: {
      type: string
      info: string
    }
  }
  /**
   * Pointer to round metadata in a decentralized storage e.g IPFS, Ceramic etc.
   */
  store?: MetadataPointer
  /**
   * Metadata of a round application to be stored off-chain
   */
  applicationMetadata?: ApplicationMetadata
  /**
   * Pointer to application metadata in a decentralized storage e.g IPFS, Ceramic etc.
   */
  applicationStore?: MetadataPointer
  /**
   * Voting contract address
   */
  votingStrategy: string
  /**
   * Payout contract address
   */
  payoutStrategy: {
    id: string
    isReadyForPayout: boolean
  }
  /**
   * Unix timestamp of the start of the round
   */
  roundStartTime: Date
  /**
   * Unix timestamp of the end of the round
   */
  roundEndTime: Date
  /**
   * Unix timestamp of when grants can apply to a round
   */
  applicationsStartTime: Date
  /**
   * Unix timestamp after which grants cannot apply to a round
   */
  applicationsEndTime: Date
  /**
   * Contract address of the token used to payout match amounts at the end of a round
   */
  token: string
  /**
   * Contract address of the program to which the round belongs
   */
  ownedBy: string
  /**
   * Addresses of wallets that will have admin privileges to operate the Grant program
   */
  operatorWallets?: Array<string>
  /**
   * List of projects approved for the round
   */
  approvedProjects?: ApprovedProject[]
  /**
   * Round fees percentage
   */
  feesPercentage?: number
  /**
   * Round fees address
   */
  feesAddress?: string

  finalized: boolean
  protocolFeePercentage?: number
  roundFeePercentage?: number
}

export type MatchingStatsData = {
  index?: number
  projectName: string
  uniqueContributorsCount?: number
  contributionsCount: number
  matchPoolPercentage: number
  projectId: string
  applicationId: string
  matchAmountInToken: BigNumber
  originalMatchAmountInToken: BigNumber
  projectPayoutAddress: string
  status?: string
  hash?: string
}

export type ProjectStatus = 'PENDING' | 'APPROVED' | 'REJECTED' | 'CANCELLED' | 'APPEAL' | 'FRAUD'

export type ProjectCredentials = {
  [key: string]: VerifiableCredential
}
interface ProjectOwner {
  address: string
}

export type ApprovedProject = {
  grantApplicationId: GrantApplicationId
  projectRegistryId: ProjectRegistryId
  recipient: recipient
  projectMetadata: ProjectMetadata
  status: ApplicationStatus
}
export type GrantApplicationId = string
export type ProjectRegistryId = string
export type recipient = string

export type ProjectMetadata = {
  title: string
  description: string
  website: string
  bannerImg?: string
  logoImg?: string
  projectTwitter?: string
  userGithub?: string
  projectGithub?: string
  credentials?: ProjectCredentials
  owners: ProjectOwner[]
}
// ...
export enum ApplicationStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  CANCELLED = 'CANCELLED',
}